myems-aggregation/requirements.txt
mysql-connector-python
python-decouple